GEM-18835: Update versions
- awaitility: 4.2.1 -> 4.2.2
- ben-manes-versions: 0.50.0 -> 0.54.0
- gemfire: 10.1.4 -> 10.1.7
- gradle: 8.14.3 -> 9.5.1
- nl.littlerobots.version-catalog-update: 0.8.5 -> 1.1.0
- spring-boot.version: 3.5.7 -> 3.5.14
- spring-data-bom.version: 2025.0.5 -> 2025.0.11
- spring-framework.version: 6.2.12 -> 6.2.18
- spring-security.version: 6.5.6 -> 6.5.10
- spring-session.version: 3.5.3 -> 3.5.6
- springBootVersion: 3.5.7 -> 3.5.14

ai-assisted=yes
