+++ gradle.properties
@@ -1,22 +1,22 @@
 #
-# Copyright 2022-2025 Broadcom. All rights reserved.
+# Copyright 2022-2026 Broadcom. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 
 org.gradle.jvmargs=-Xmx2g -Dfile.encoding=UTF-8
 version=1.0.0-build.9999
 
-gemfireVersion=10.1.4
+gemfireVersion=10.1.7
 springSessionDataGemFireVersion=1.0.0
 springDataGemFireVersion=1.0.0
 
 additionalMavenRepoURLs=''
 
-spring-framework.version=6.2.12
-spring-data-bom.version=2025.0.5
-spring-boot.version=3.5.7
-spring-session.version=3.5.3
-spring-security.version=6.5.6
+spring-framework.version=6.2.18
+spring-data-bom.version=2025.0.11
+spring-boot.version=3.5.14
+spring-session.version=3.5.6
+spring-security.version=6.5.10
 
 org.gradle.workers.max=3
 org.gradle.parallel=true
