+++ gradle/libs.versions.toml
@@ -1,48 +1,46 @@
 [versions]
 antlrVersion = "2.7.7"
-awaitilityVersion="4.2.1"
 aspectjToolsVersion = "1.9.22.1"
+awaitilityVersion = "4.2.2"
+benManesVersions = "0.54.0"
 findbugsVersion = "3.0.2"
 gemFireTestContainersVersion = "2.3.3"
+gemfireVersion = "10.1.7"
+littlerobotsVersionCatalogUpdate = "1.1.0"
+lombokPlugin = "8.6"
 mockitoCoreVersion = { strictly = "5.14.2" }
 multithreadedtcVersion = "1.01"
-springBootVersion = "3.5.7"
+springBootVersion = "3.5.14"
+springDataGemFireVersion = "1.0.0"
+springSessionDataGemFireVersion = "1.0.0"
 springShellVersion = "1.2.0.RELEASE"
-
-benManesVersions = "0.50.0"
-littlerobotsVersionCatalogUpdate = "0.8.5"
-lombokPlugin = "8.6"
 
 [libraries]
 antlr = { module = "antlr:antlr", version.ref = "antlrVersion" }
 aspectj-tools = { module = "org.aspectj:aspectjtools", version.ref = "aspectjToolsVersion" }
-awaitility={module="org.awaitility:awaitility",version.ref="awaitilityVersion"}
+awaitility = { module = "org.awaitility:awaitility", version.ref = "awaitilityVersion" }
 findbugs-jsr305 = { module = "com.google.code.findbugs:jsr305", version.ref = "findbugsVersion" }
-gemfire-testcontainers = { module = "dev.gemfire:gemfire-testcontainers", version.ref = "gemFireTestContainersVersion" }
-multithreadedtc = { module = "edu.umd.cs.mtc:multithreadedtc", version.ref = "multithreadedtcVersion" }
-mockito-core = { module = "org.mockito:mockito-core", version.ref = "mockitoCoreVersion" }
-spring-boot-configuration-processor = { module = "org.springframework.boot:spring-boot-configuration-processor", version.ref = "springBootVersion" }
-spring-boot-autoconfigure-processor = { module = "org.springframework.boot:spring-boot-autoconfigure-processor", version.ref = "springBootVersion" }
-spring-shell = { module = "org.springframework.shell:spring-shell", version.ref = "springShellVersion" }
-
 gemfire-core = { module = "com.vmware.gemfire:gemfire-core", version.ref = "gemfireVersion" }
-gemfire-gfsh = { module = "com.vmware.gemfire:gemfire-gfsh", version.ref = "gemfireVersion" }
-gemfire-web = { module = "com.vmware.gemfire:gemfire-web", version.ref = "gemfireVersion" }
-gemfire-logging = { module = "com.vmware.gemfire:gemfire-logging", version.ref = "gemfireVersion" }
 gemfire-cq = { module = "com.vmware.gemfire:gemfire-cq", version.ref = "gemfireVersion" }
-gemfire-wan = { module = "com.vmware.gemfire:gemfire-wan", version.ref = "gemfireVersion" }
-gemfire-membership = { module = "com.vmware.gemfire:gemfire-membership", version.ref = "gemfireVersion" }
-gemfire-tcp-server = { module = "com.vmware.gemfire:gemfire-tcp-server", version.ref = "gemfireVersion" }
-gemfire-serialization = { module = "com.vmware.gemfire:gemfire-serialization", version.ref = "gemfireVersion" }
 gemfire-deployment-chained-classloader = { module = "com.vmware.gemfire:gemfire-deployment-chained-classloader", version.ref = "gemfireVersion" }
-
+gemfire-gfsh = { module = "com.vmware.gemfire:gemfire-gfsh", version.ref = "gemfireVersion" }
+gemfire-logging = { module = "com.vmware.gemfire:gemfire-logging", version.ref = "gemfireVersion" }
+gemfire-membership = { module = "com.vmware.gemfire:gemfire-membership", version.ref = "gemfireVersion" }
+gemfire-serialization = { module = "com.vmware.gemfire:gemfire-serialization", version.ref = "gemfireVersion" }
+gemfire-tcp-server = { module = "com.vmware.gemfire:gemfire-tcp-server", version.ref = "gemfireVersion" }
+gemfire-testcontainers = { module = "dev.gemfire:gemfire-testcontainers", version.ref = "gemFireTestContainersVersion" }
+gemfire-wan = { module = "com.vmware.gemfire:gemfire-wan", version.ref = "gemfireVersion" }
+gemfire-web = { module = "com.vmware.gemfire:gemfire-web", version.ref = "gemfireVersion" }
+mockito-core = { module = "org.mockito:mockito-core", version.ref = "mockitoCoreVersion" }
+multithreadedtc = { module = "edu.umd.cs.mtc:multithreadedtc", version.ref = "multithreadedtcVersion" }
+spring-boot-autoconfigure-processor = { module = "org.springframework.boot:spring-boot-autoconfigure-processor", version.ref = "springBootVersion" }
+spring-boot-configuration-processor = { module = "org.springframework.boot:spring-boot-configuration-processor", version.ref = "springBootVersion" }
 spring-data-gemfire = { module = "com.vmware.gemfire:spring-data-3.5-gemfire-10.1", version.ref = "springDataGemFireVersion" }
 spring-data-gemfire-test-framework = { module = "com.vmware.gemfire:spring-data-3.5-gemfire-test-framework-10.1", version.ref = "springDataGemFireVersion" }
 spring-session-gemfire = { module = "com.vmware.gemfire:spring-session-3.5-gemfire-10.1", version.ref = "springSessionDataGemFireVersion" }
-
-[bundles]
+spring-shell = { module = "org.springframework.shell:spring-shell", version.ref = "springShellVersion" }
 
 [plugins]
-lombok = { id = "io.freefair.lombok", version.ref = "lombokPlugin" }
 ben-manes-versions = { id = "com.github.ben-manes.versions", version.ref = "benManesVersions" }
 littlerobots-version-catalog-update = { id = "nl.littlerobots.version-catalog-update", version.ref = "littlerobotsVersionCatalogUpdate" }
+lombok = { id = "io.freefair.lombok", version.ref = "lombokPlugin" }
